Fix recursion bug
nextNum() was using i from preterNums.map() as the index, but
i doesn't increment until the next iteration of map(), which
doesn't occur until all recursive calls to nextNum() resolve;
this resulted in an infinite recursion loop. Passing the index
to nextNum as an argument allowed it to increment that for its
next recursion call, and now has the expected behavior.